Froomtened backened integration almost done

+++ lib/getX_controllers/detail-screen/heart_rate_controllers.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_rx/get_rx.dart';
 import 'package:get/get_rx/src/rx_types/rx_types.dart';
+import 'package:get/state_manager.dart';
 import 'package:health_guardian/styling/sizeConfig.dart';
 import 'package:health_guardian/styling/toast_message.dart';
 import 'package:intl/intl.dart';
@@ -41,16 +42,16 @@ List<dynamic> getHeartColorPosition(num heartBPMRate) {
 
   if (heartBPMRate < 60) {
     color = Colors.blue;
-    position = 7.8125*SizeConfig.widthMultiplier;
+    position = 7.8125 * SizeConfig.widthMultiplier;
   } else if (isBetween(heartBPMRate, 60, 100)) {
     color = Colors.green;
-    position = 40.178571*SizeConfig.widthMultiplier;
+    position = 40.178571 * SizeConfig.widthMultiplier;
   } else if (heartBPMRate > 100) {
     color = Colors.red;
-    position = 69.19642*SizeConfig.widthMultiplier;
+    position = 69.19642 * SizeConfig.widthMultiplier;
   } else {
     color = Colors.grey;
-    position = 7.8125*SizeConfig.widthMultiplier;
+    position = 7.8125 * SizeConfig.widthMultiplier;
   }
 
   return [color, position];
@@ -59,6 +60,7 @@ List<dynamic> getHeartColorPosition(num heartBPMRate) {
 class HeartRateControllers extends GetxController {
   RxDouble heartRate = 0.0.obs;
   Random random = Random();
+  RxBool isTwoDigit = false.obs;
   RxBool animationBool = false.obs;
   int animationCounter = 0;
 
@@ -74,6 +76,7 @@ class HeartRateControllers extends GetxController {
 
   void getHeartRate() {
     heartRate.value = 70.0 + random.nextInt(71);
+    heartBPMRate.value = heartRate.value;
   }
 
   void reCheck() {
@@ -92,9 +95,9 @@ class HeartRateControllers extends GetxController {
   RxDouble arrowPosition = 35.0.obs;
 
   //* Heart Rate Value entered by the user
-  RxInt heartBPMRate = 0.obs;
+  RxDouble heartBPMRate = 0.0.obs;
 
-  void changeLevel(int value) {
+  void changeLevel(double value) {
     heartBPMRate.value = value;
     List<dynamic> list = getHeartColorPosition(heartBPMRate.value);
     arrowColor.value = list[0];
@@ -154,9 +157,14 @@ class HeartRateControllers extends GetxController {
           FirebaseFirestore.instance.collection("heart_data");
       QuerySnapshot querySnapshot =
           await collectionReference.where("email", isEqualTo: email).get();
-      DocumentSnapshot docs = await querySnapshot.docs.first;
 
-      if (!docs.exists) {
+      if (heartBPMRate.value >= 10 && heartBPMRate.value < 100) {
+        heartBPMRate.value =
+            double.parse((heartBPMRate.value).toStringAsFixed(2));
+        isTwoDigit.value = true;
+      }
+
+      if (querySnapshot.docs.isEmpty) {
         await collectionReference.add({
           "email": email,
           "heart_data": [
@@ -166,17 +174,20 @@ class HeartRateControllers extends GetxController {
               "color": color,
               "status": status,
               "state": State.value,
+              "isTwoDigit": isTwoDigit.value,
               "note": noteController.text.toString()
             }
           ]
         });
       } else {
+        DocumentSnapshot docs = await querySnapshot.docs.first;
         final list = {
           "date": "${date.toString()} : ${time.toString()}",
           "heart_level": heartBPMRate.value,
           "color": color,
           "status": status,
           "state": State.value,
+          "isTwoDigit": isTwoDigit.value,
           "note": noteController.text.toString()
         };
 
@@ -196,8 +207,15 @@ class HeartRateControllers extends GetxController {
 }
 
 class EditHeartRateDataController extends GetxController {
+  @override
+  void onInit() {
+    super.onInit();
+    fetchHeartData();
+  }
+
   RxList heart_data_list = [].obs;
   RxList heart_graph_list = [].obs;
+  RxList heart_report_list = [].obs;
 
   RxBool isLoadingReport = false.obs;
 
@@ -208,6 +226,7 @@ class EditHeartRateDataController extends GetxController {
         toastErrorSlide(context, "No Data To Store");
         return;
       }
+
       isLoadingReport.value = true;
 
       Map<String, List<double>> heartBPMData = {};
@@ -216,8 +235,8 @@ class EditHeartRateDataController extends GetxController {
       final date = DateFormat('dd MMM yyyy').format(DateTime.now());
 
       for (var data in heart_data_list) {
-        String date = data["date"];
-        double heartLevel = data["heart_level"];
+        String date = (data["date"] as String).split(":")[0];
+        double heartLevel = (data["heart_level"] as num).toDouble();
 
         //* Add sygar values
         if (!heartBPMData.containsKey(date)) {
@@ -239,6 +258,7 @@ class EditHeartRateDataController extends GetxController {
         final colorStatusList = getHeartColorStatus(avgHeartLevel[key]!);
         heart_report.add({
           "heart_level": avgHeartLevel[key],
+          "date": key,
           "color": colorStatusList[0],
           "status": colorStatusList[1]
         });
@@ -249,9 +269,8 @@ class EditHeartRateDataController extends GetxController {
           FirebaseFirestore.instance.collection("heart_report");
       QuerySnapshot querySnapshot =
           await collectionReference.where('email', isEqualTo: email).get();
-      DocumentSnapshot docs = querySnapshot.docs.first;
 
-      if (!docs.exists) {
+      if (querySnapshot.docs.isEmpty) {
         await collectionReference.add({
           "email": email,
           "heart_report": [
@@ -259,66 +278,83 @@ class EditHeartRateDataController extends GetxController {
           ]
         });
       } else {
+        DocumentSnapshot docs = querySnapshot.docs.first;
         final list = {"submitted_on": date, "heart_report": heart_report};
 
         await docs.reference.update({
           "heart_report": FieldValue.arrayUnion([list])
         });
+      }
 
-        //* after report is stored clear the data
-        CollectionReference collectionReference =
-            FirebaseFirestore.instance.collection("heart_data");
-        QuerySnapshot querySnapshot =
-            await collectionReference.where('email', isEqualTo: email).get();
-        DocumentSnapshot docs1 = querySnapshot.docs.first;
+      //* Clear heart_data if exists
+      CollectionReference heartDataCollection =
+          FirebaseFirestore.instance.collection("heart_data");
+      QuerySnapshot heartDataSnapshot =
+          await heartDataCollection.where('email', isEqualTo: email).get();
+
+      if (heartDataSnapshot.docs.isNotEmpty) {
+        DocumentSnapshot docs1 = heartDataSnapshot.docs.first;
         await docs1.reference.update({
           'heart_data': FieldValue.delete(),
         });
-
-        toastSuccessSlide(context, "Report Stored Successfully!");
       }
+      toastSuccessSlide(context, "Report Generate Successfully!");
 
       isLoadingReport.value = false;
     } catch (e) {
+      toastErrorSlide(context, e.toString());
       isLoadingReport.value = false;
     }
   }
 
-  //* fetch heart data from database
-  Future<List<dynamic>> fetchHeartData(BuildContext context) async {
+  //* fetch heart data stream from database
+  Stream<List<dynamic>> fetchHeartData() async* {
     try {
       final email = FirebaseAuth.instance.currentUser!.email!;
 
       CollectionReference collectionReference =
           FirebaseFirestore.instance.collection("heart_data");
-      QuerySnapshot querySnapshot =
-          await collectionReference.where('email', isEqualTo: email).get();
-      DocumentSnapshot docs = querySnapshot.docs.first;
 
-      List<dynamic> list = docs['heart_data'].toList();
-      heart_data_list.value = list;
-
-      return heart_data_list;
+      yield* collectionReference
+          .where('email', isEqualTo: email)
+          .snapshots()
+          .map((querySnapshot) {
+        if (querySnapshot.docs.isNotEmpty) {
+          DocumentSnapshot docs = querySnapshot.docs.first;
+          List<dynamic> list = docs['heart_data'].toList();
+          heart_data_list.value = list;
+          return list;
+        } else {
+          return [];
+        }
+      });
     } catch (e) {
-      return [];
+      yield [];
     }
   }
 
   //* get report of heart data
-  Future<List<dynamic>> getHeartReportData() async {
+  Stream<List<dynamic>> getHeartReportData() {
     try {
       final email = FirebaseAuth.instance.currentUser!.email!;
 
       CollectionReference collectionReference =
           FirebaseFirestore.instance.collection("heart_report");
-      QuerySnapshot querySnapshot =
-          await collectionReference.where('email', isEqualTo: email).get();
-      DocumentSnapshot docs = querySnapshot.docs.first;
 
-      final List<dynamic> list = docs['heart_report'];
-      return list;
+      return collectionReference
+          .where('email', isEqualTo: email)
+          .snapshots()
+          .map((querySnapshot) {
+        if (querySnapshot.docs.isNotEmpty) {
+          DocumentSnapshot docs = querySnapshot.docs.first;
+          heart_report_list.value = docs['heart_report'];
+          return docs['heart_report'] as List<dynamic>;
+        } else {
+          return [];
+        }
+      });
     } catch (e) {
-      return [];
+      return Stream.value([]);
     }
   }
 }
